ByOperations

Documentation for ByOperations.

ByOperations.By

ByOperations.By — Method

By(f=identity)
By{K,V}(f=identity)

By creates an object that triggers "key-selective" operations on a collection. f is the function that generates the key, and the operation determines the resulting value. The return value is typically a Dict. Optionally, you can specify the key K and value V types of that Dict, which can help performance in certain cases (see the documentation for details).

Examples

julia> count(By(lowercase), "Hello")
Dict{Char, Int64} with 4 entries:
  'h' => 1
  'l' => 2
  'e' => 1
  'o' => 1

julia> push!(By(isodd), 1:11)
Dict{Bool, Vector{Int64}} with 2 entries:
  0 => [2, 4, 6, 8, 10]
  1 => [1, 3, 5, 7, 9, 11]